Add support for GCAFS on MSU HPC (#3844)
This adds support for GCAFS cycling on MSU HPC (that do not have
access to HPSS) by allowing for the extraction of GDAS files from
prestaged tarballs on disk. This includes a shell script to run on
Mercury/Ursa/other machines to prepare the files that will then need to
be copied to the MSU HPC disks.