Move j-jobs and ex-scripts into dev subdirectory (#4259)
This moves `jobs/` to `dev/jobs`, `dev/jobs` to
`dev/job_cards/rocoto` and `scripts` to `dev/scripts` (but symlinks the
latter to the original location for now. This is in anticipation for T2O
where we will only want to put in `jobs/` and `scripts/` the files
needed for each application from the total global-workflow develop set.